Quick hack to support R5

+++ src/include/crc32.h
@@ -21,6 +21,6 @@
 #ifndef __CRC32_H
 #define __CRC32_H
 
-uint32_t generic_crc32(target *t, uint32_t base, int len);
+uint32_t generic_crc32(target *t, uint32_t base, size_t len);
 
 #endif
